+++ CoDrone/__init__.py
@@ -1,8 +1,14 @@
+from CoDrone.codrone import CoDrone
+from CoDrone.codrone import *
+
 __all__ = [
     "crc",
-    "codronee.py",
+    "codrone",
+    "codrone.CoDrone",
     "protocol",
     "receiver",
     "storage",
     "system",
     ]
+
+
