+++ CoDrone/receiver.py
@@ -1,8 +1,9 @@
-import time
+from enum import Enum
 from crc import CRC16
-from protocol import *
+from time import clock
 
-
+from protocol import Header
+from protocol import DataType
 
 class StateLoading(Enum):
     Ready = 0x00
@@ -40,7 +41,7 @@ class Receiver:
 
     def call(self, data):
 
-        now = time.clock() * 1000
+        now = clock() * 1000
 
         self.message = None
