+++ setup.py
@@ -14,12 +14,13 @@ this is Python package for control Codrone
 
 setup(
     name='CoDrone',
-    version='0.2',
+    version='1.0',
     description='Python package for CoDrone',
     url='https://github.com/RobolinkInc/CoDrone-python.git',
     author='robolink',
     author_email='whoseop@robolink.com',
     packages=["CoDrone"],
+    keywords=['robolink','drone'],
     include_package_data=True,
     install_requires=install_requires,
     setup_requires=setup_requires,
